+++ cmd/task.go
@@ -9,7 +9,7 @@ var taskCmd = &cobra.Command{
 	Aliases: []string{"t"},
 	Short:   "Group of commands related to `task`.",
 	Long: `This set of commands relate to Robocorp Control Room related tasks. They are
-executed either locally, or in connection to Robocorp Control Room and Robocorp App.`,
+executed either locally, or in connection to Robocorp Control Room and tooling.`,
 }
 
 func init() {